Revert "Add pipeines func tests (#70)"
This reverts commit f2f74ba709b9dad7224de49ae5c46cb1d8759d86.
