--- src/test/java/it/com/atlassian/bitbucket/jenkins/internal/scm/BitbucketSCMStepDescriptorIT.java
@@ -1,9 +1,0 @@
-package it.com.atlassian.bitbucket.jenkins.internal.scm;
-
-public class BitbucketSCMStepDescriptorIT extends AbstractBitbucketSCMDescriptorIT {
-
-    @Override
-    protected String getClassName() {
-        return "BitbucketSCMStep";
-    }
-}
